Refactor forms and remove deprecated calls
