+++ tests/views/test_reviewers_management.py
@@ -28,21 +28,19 @@ import datetime
 from django.forms import formset_factory
 from django.test import RequestFactory, TestCase, Client
 
-from base.models.entity import find_versions_from_entites
-from base.models.enums import entity_type
-from base.tests.factories.academic_year import AcademicYearFactory
-from base.tests.factories.entity import EntityFactory
-from base.tests.factories.entity_version import EntityVersionFactory
-from base.tests.factories.person import PersonFactory
-
-from assistant.forms import ReviewersFormset
+from assistant.forms.reviewer import ReviewersFormset
 from assistant.models.enums import reviewer_role
 from assistant.tests.factories.academic_assistant import AcademicAssistantFactory
 from assistant.tests.factories.assistant_mandate import AssistantMandateFactory
 from assistant.tests.factories.manager import ManagerFactory
 from assistant.tests.factories.review import ReviewFactory
 from assistant.tests.factories.reviewer import ReviewerFactory
-
+from base.models.entity import find_versions_from_entites
+from base.models.enums import entity_type
+from base.tests.factories.academic_year import AcademicYearFactory
+from base.tests.factories.entity import EntityFactory
+from base.tests.factories.entity_version import EntityVersionFactory
+from base.tests.factories.person import PersonFactory
 
 HTTP_OK = 200
 HTTP_FOUND = 302
