+++ views/phd_supervisor_review.py
@@ -25,17 +25,15 @@
 ##############################################################################
 from django.contrib.auth.decorators import user_passes_test
 from django.core.exceptions import ObjectDoesNotExist
-from django.urls import reverse
 from django.http.response import HttpResponseRedirect
 from django.shortcuts import render
+from django.urls import reverse
 from django.utils import timezone
 from django.views.decorators.http import require_http_methods
 
-from base.models.enums import entity_type
-
 from assistant.business.mandate_entity import get_entities_for_mandate
 from assistant.business.users_access import user_is_phd_supervisor_and_procedure_is_open
-from assistant.forms import ReviewForm
+from assistant.forms.review import ReviewForm
 from assistant.models import assistant_document_file
 from assistant.models import assistant_mandate
 from assistant.models import mandate_entity
@@ -46,6 +44,7 @@ from assistant.models.enums import assistant_mandate_state
 from assistant.models.enums import document_type
 from assistant.models.enums import review_status
 from assistant.models.enums import reviewer_role
+from base.models.enums import entity_type
 
 
 @require_http_methods(["POST"])
