+++ views/reviewers_management.py
@@ -25,21 +25,21 @@
 ##############################################################################
 from django.contrib.auth.decorators import user_passes_test
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
-from django.urls import reverse
 from django.forms.forms import NON_FIELD_ERRORS
 from django.forms.formsets import formset_factory
 from django.http import HttpResponseRedirect
 from django.shortcuts import render, redirect
+from django.urls import reverse
 from django.utils.translation import ugettext as _
 from django.views.decorators.http import require_http_methods
 from django.views.generic import ListView
 from django.views.generic.edit import FormMixin
 
-from base.models import academic_year, person, entity_version
-
-from assistant.forms import MandatesArchivesForm, ReviewersFormset, ReviewerForm, ReviewerReplacementForm
+from assistant.forms.mandate import MandatesArchivesForm
+from assistant.forms.reviewer import ReviewerForm, ReviewerReplacementForm, ReviewersFormset
 from assistant.models import reviewer, review
 from assistant.utils import manager_access
+from base.models import academic_year, person, entity_version
 
 
 class ReviewersListView(LoginRequiredMixin, UserPassesTestMixin, ListView, FormMixin):
